requirements.txt
certifi==2020.4.5.1
clickhouse-driver==0.1.3
dataclasses==0.6; python_version < "3.7"
kafka-python==2.0.1
mysql-replication==0.21
mysqlclient==1.4.6
pydantic==1.5.1
pymysql==0.9.3
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
sentry-sdk==0.14.4
six==1.14.0
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.9
